+++ frontend/src/AuthenticatedApp/components/Header/AvatarMenu/index.js
@@ -1,7 +1,7 @@
 import { Avatar, Box, IconButton, Menu, MenuItem } from "@mui/material"
 import { useState } from "react";
-import { useAuthContext } from "../../../hooks/auth/useAuthContext";
-import WarningDialog from "../../WarningDialog";
+import { useAuthContext } from "../../../../hooks/auth/useAuthContext";
+import WarningDialog from "../../../../components/WarningDialog";
 import ChangePasswordModal from "./ChangePasswordModal";
 import DeleteAccountModal from "./DeleteAccountModal";
 import LogOutButton from "./LogOutButton";
